reorganising component code
