+++ src/components/modals/shortlist-it-add-criteria-from-template-modal.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect } from "react";
-import { ShortlistItStateManager } from "../types/shortlist-it-state-manager";
-import ShortlistItCriteriaTemplateItem from "./shortlist-it-criteria-template-item";
+import { ShortlistItStateManager } from "../../types/shortlist-it-state-manager";
+import ShortlistItCriteriaTemplateItem from "../shortlist-it-criteria-template-item";
 import { ShortlistItModal } from "./shortlist-it-modal";
-import { getList } from "../component-actions/list-actions";
+import { getList } from "../../component-actions/list-actions";
 import { ListGroup } from "react-bootstrap";
 
 type ShortlistItAddCriteriaFromTemplateModalProps = {
