+++ src/components/modals/shortlist-it-criteria-deletion-modal.tsx
@@ -1,9 +1,9 @@
 import React from "react";
-import { ShortlistItStateManager } from "../types/shortlist-it-state-manager";
+import { ShortlistItStateManager } from "../../types/shortlist-it-state-manager";
 import { ShortlistItModal } from "./shortlist-it-modal";
 import { Button } from "react-bootstrap";
-import { store } from "../utilities/storage";
-import { Shortlist } from "../types/shortlist";
+import { store } from "../../utilities/storage";
+import { Shortlist } from "../../types/shortlist";
 
 type ShortlistItCriteriaDeletionModalProps = {
     stateMgr: ShortlistItStateManager;
