+++ src/components/modals/shortlist-it-list-deletion-modal.tsx
@@ -1,9 +1,9 @@
 import React from "react";
-import { ShortlistItStateManager } from "../types/shortlist-it-state-manager";
+import { ShortlistItStateManager } from "../../types/shortlist-it-state-manager";
 import { ShortlistItModal } from "./shortlist-it-modal";
 import { Button } from "react-bootstrap";
-import { store } from "../utilities/storage";
-import { archiveList } from "../component-actions/list-actions";
+import { store } from "../../utilities/storage";
+import { archiveList } from "../../component-actions/list-actions";
 
 type ShortlistItListDeletionModalProps = {
     stateMgr: ShortlistItStateManager;
